Use enums and split KDF logic into its own file
